Added ost of classes needed

+++ RTS Project code/Motor2D/Render.cpp
@@ -223,7 +223,7 @@ iPoint Render::ScreenToWorld(int x, int y) const
 bool Render::Blit(SDL_Texture* texture, int x, int y, const SDL_Rect* section, bool flip, float speed, double angle, int pivot_x, int pivot_y) const
 {
 	bool ret = true;
-	uint scale = App->win->GetScale();
+	float scale = App->win->GetScale();
 
 	SDL_Rect rect;
